+++ src/collective/cover/tests/test_setup.py
@@ -10,8 +10,10 @@ import unittest
 
 JS = [
     '++resource++collective.cover/js/contentchooser.js',
-    '++resource++collective.js.bootstrap/js/bootstrap.min.js',
+    '++resource++collective.cover/js/layout_edit.js',
+    '++resource++collective.cover/js/main.js',
     '++resource++collective.cover/js/vendor/jquery.endless-scroll.js',
+    '++resource++collective.js.bootstrap/js/bootstrap.min.js',
 ]
 
 CSS = [
@@ -60,6 +62,17 @@ class InstallTestCase(unittest.TestCase):
         except AttributeError:
             self.fail('Reinstall fails when the record was changed')
 
+    def test_versioning_policy(self):
+        repository = self.portal['portal_repository']
+        policy_map = repository.getPolicyMap()['collective.cover.content']
+        self.assertEqual(policy_map, [u'version_on_revert'])
+
+    @unittest.skipIf(IS_PLONE_5, 'Plone 4.3 only')
+    def test_tinymce_linkable(self):
+        tinymce = self.portal['portal_tinymce']
+        linkable = tinymce.linkable.split('\n')
+        self.assertIn('collective.cover.content', linkable)
+
 
 class UninstallTestCase(unittest.TestCase):
 
@@ -90,3 +103,15 @@ class UninstallTestCase(unittest.TestCase):
         resource_ids = self.portal.portal_css.getResourceIds()
         for id in CSS:
             self.assertNotIn(id, resource_ids, '{0} not removed'.format(id))
+
+    @unittest.expectedFailure  # XXX: not pretty sure how to test this
+    def test_versioning_policy_removed(self):
+        repository = self.portal['portal_repository']
+        policy_map = repository.getPolicyMap()
+        self.assertNotIn('collective.cover.content', policy_map)
+
+    @unittest.skipIf(IS_PLONE_5, 'Plone 4.3 only')
+    def test_tinymce_linkable_removed(self):
+        tinymce = self.portal['portal_tinymce']
+        linkable = tinymce.linkable.split('\n')
+        self.assertNotIn('collective.cover.content', linkable)